表をコンポーネント化中
