+++ resources/js/Components/Molecules/TableData.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const TableData = () => {
+    return <div>TableData</div>
+}
+
+export default TableData
